fix cd norm
