+++ srcs/buildin/cd.c
@@ -6,33 +6,11 @@
 /*   By: marai <marai@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/15 12:59:56 by keys              #+#    #+#             */
-/*   Updated: 2023/06/24 19:49:42 by marai            ###   ########.fr       */
+/*   Updated: 2023/06/24 22:28:31 by marai            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-#define PATH_MAXLEN 4096
-
-void	imple_pwd(t_env **head, t_status *s);
-
-char	*get_home_dir(t_env *env)
-{
-	char	*homepath;
-
-	while (env)
-	{
-		if (!ft_strcmp(env->key, "HOME"))
-		{
-			homepath = ft_strdup(env->value);
-			if (!homepath)
-				_err_malloc();
-			return (homepath);
-		}
-		env = env->next;
-	}
-	return (NULL);
-}
 
 char	*make_abs_path(char *path, char *argv, char *home)
 {
@@ -40,23 +18,11 @@ char	*make_abs_path(char *path, char *argv, char *home)
 
 	i = 0;
 	if (!ft_strcmp(argv, "~"))
-	{
-		if (!home)
-		{
-			ft_putendl_fd("HOME not set", STDERR_FILENO);
-			return (NULL);
-		}
-		ft_memset(path, '\0', PATH_MAXLEN);
-		ft_strlcpy(path, home, PATH_MAXLEN);
-		return (path);
-	}
+		return (make_home_path(path, home));
 	if (!ft_strncmp(argv, "~/", 2))
 	{
-		if (!home)
-		{
-			ft_putendl_fd("HOME not set", STDERR_FILENO);
+		if (!is_home_set(home))
 			return (NULL);
-		}
 		ft_memset(path, '\0', PATH_MAXLEN);
 		ft_strlcpy(path, home, PATH_MAXLEN);
 		i = 2;
@@ -76,14 +42,13 @@ int	move_to_abs_path(char *path)
 {
 	int		status;
 
-	status =  chdir(path);
+	status = chdir(path);
 	if (status < 0)
 		ft_putendl_fd(" No such file or directory", STDERR_FILENO);
-	
 	return (status);
 }
 
-int	move_to_home_or_rel_path(char *path[], t_env *env, t_status *s)
+int	move_to_home_or_rel_path(char *path[], t_env *env)
 {
 	char	home_path[PATH_MAXLEN];
 	char	*home;
@@ -92,12 +57,8 @@ int	move_to_home_or_rel_path(char *path[], t_env *env, t_status *s)
 	home = get_home_dir(env);
 	if (!path[1])
 	{
-		if (!home)
-		{
-			ft_putendl_fd("HOME not set", STDERR_FILENO);
-			s->status = 1;
+		if (!is_home_set(home))
 			return (-1);
-		}
 		status = chdir(home);
 	}
 	else
@@ -111,8 +72,7 @@ int	move_to_home_or_rel_path(char *path[], t_env *env, t_status *s)
 		free(home);
 	if (status < 0)
 		ft_putendl_fd(" No such file or directory", STDERR_FILENO);
-	
-	return status;
+	return (status);
 }
 
 void	handle_cd_status(int status, t_env **env, t_status *s)
@@ -133,78 +93,15 @@ int	cd(char *argv[], t_env **env, t_status *s)
 		ft_putendl_fd("minishell: cd: too many arguments", STDERR_FILENO);
 		status = -1;
 	}
+	else if (argv[1] && PATH_MAXLEN - 1 < ft_strlen(argv[1]))
+	{
+		ft_putendl_fd("minishell: cd: too long path", STDERR_FILENO);
+		status = -1;
+	}
 	else if (argv[1] && argv[1][0] == '/')
 		status = move_to_abs_path(argv[1]);
 	else
-		status = move_to_home_or_rel_path(argv, *env, s);
+		status = move_to_home_or_rel_path(argv, *env);
 	handle_cd_status(status, env, s);
 	return (0);
-}
-
-
-// int	cd(char *argv[], t_env *env, t_status *s)
-// {
-// 	char	path[PATH_MAXLEN];
-// 	char	*home;
-// 	int		status;
-
-// 	s->f = true;
-// 	if (argv[1][0] == '/')
-// 		status = chdir(argv[1]);
-// 	else
-// 	{
-// 		home = get_home_dir(env);
-// 		if (home == NULL)
-// 		{
-// 			ft_putendl_fd("HOME not set", STDERR_FILENO);
-// 			s->status = 1;
-// 			return (-1);
-// 		}
-// 		if (!argv[1])
-// 			status = chdir(home);
-// 		else
-// 		{
-// 			ft_memset(path, '\0', PATH_MAXLEN);
-// 			getcwd(path, PATH_MAXLEN);
-// 			make_abs_path(path, argv[1], home);
-// 			status = chdir(path);
-// 			free(home);
-// 		}
-// 	}
-// 	if (status < 0)
-// 	{
-// 		ft_putendl_fd("minishell: cd: too many arguments", STDERR_FILENO);
-// 	}
-// 	else
-// 		imple_pwd(env, s);
-// 	return (0);
-// }
-
-void	imple_pwd(t_env **head, t_status *s)
-{
-	char	*path;
-	bool	is_pwd;
-	t_env	*env;
-
-	env = *head;
-	is_pwd = false;
-	while (env)
-	{
-		if (!ft_strcmp("PWD", env->key))
-		{
-			if (env->value)
-			{
-				ft_env_addback(head, make_env("OLDPWD", env->value));
-				is_pwd = true;
-			}
-		}
-		env = env->next;
-	}
-	if (!is_pwd && head)
-		ft_env_addback(head, make_env("OLDPWD", NULL));
-	path = get_pwd(s);
-	if (head && path)
-		ft_env_addback(head, make_env("PWD", path));
-	if (path)
-		free(path);
 }
